stages/oci-archive: include limited set of xattrs
Only include a very specific set of extended attributes:
  - user.*: user specified extended attributes
  - security.ima: Integrity Measurement Architecture (IMA)
  - security.capability: Linux capabilities(7)
This follows what containers/storage[1] and containers/buildah[2]
are doing. It is important to note that we DO NOT want selinux
related extended attributes (`security.selinux`) in there, which
seems to be pulled in by some versions of `tar` even if that was
seemingly excluded via `--no-selinux`. Therefore we also exclude
selinux and xattrs explicitly from the wrapping container to
make sure they are never included.

[1] https://github.com/containers/storage/blob/35ebda8ae27dc130c18517cd1e14ad85407ae22f/pkg/archive/archive.go#L399
[2] https://github.com/containers/buildah/blob/214e4c93350563cef06aa713a4bd6fa6ed35c636/copier/xattrs.go#L19