inputs: make inputs aware of their names
The name of the input here refers to its id within the manifest. This
is unique per stage and this identifies a input for a given stage.
Include that in the input id calculation since this will ensure that
two inputs of the same time, with the same options will still have
different ids, since their name must be different.
Adapt the rpm stage test for the id change.